Move hook-related types from cattrs.dispatch to cattrs.types to avoid circular import from cattrs.fns (type-annotated function will be added there)
