+++ src/cattrs/types.py
@@ -1,9 +1,29 @@
 from typing import Protocol, TypeVar
 
 __all__ = ["SimpleStructureHook"]
+__all__ = [
+    "Hook",
+    "HookFactory",
+    "SimpleStructureHook",
+    "StructuredValue",
+    "StructureHook",
+    "TargetType",
+    "UnstructuredValue",
+    "UnstructureHook",
+]
 
 In = TypeVar("In")
 T = TypeVar("T")
+
+TargetType: TypeAlias = Any
+UnstructuredValue: TypeAlias = Any
+StructuredValue: TypeAlias = Any
+
+StructureHook: TypeAlias = Callable[[UnstructuredValue, TargetType], StructuredValue]
+UnstructureHook: TypeAlias = Callable[[StructuredValue], UnstructuredValue]
+
+Hook = TypeVar("Hook", StructureHook, UnstructureHook)
+HookFactory: TypeAlias = Callable[[TargetType], Hook]
 
 
 class SimpleStructureHook(Protocol[In, T]):
